refactor: move useInput to hooks folder
