+++ online-store/src/components/Busket/CheckoutModal/CheckoutModal.tsx
@@ -1,29 +1,9 @@
-import { useState, FocusEvent } from 'react';
+import { useState } from 'react';
 import './CheckoutModal.scss';
 import { getPaymentSystemIcon } from './getPaymentSystemIcon';
 import { useNavigate } from "react-router-dom";
 import { clearCart } from '../../../controllers/clearCart';
 import useValidation from '../../../hooks/useValidation';
-
-const useInput = (initialValue: string, validations: { [key: string]: boolean }) => {
-  const [value, setValue] = useState(initialValue);
-  const [isDirty, setDirty] = useState(false);
-  const valid = useValidation(value, validations);
-  const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setValue(event.target.value);
-  };
-  const onBlur = () => {
-    setDirty(true);
-  };
-  
-  return {
-    value,
-    onChange,
-    onBlur,
-    isDirty,
-    ...valid
-  }
-};
 
 function CheckoutModal(
   {
